Cleanup

Former-commit-id: 3f23f0ee20adbec20111ed25e9db8f3e7f58ea15

+++ src/librapid/python/generators/mpfrInterface.py
@@ -71,7 +71,7 @@ for type in types:
 		Function("__int__", [Argument(constRef, "this_")], "return librapid::internal::traits<librapid::{}>::cast<int64_t>(this_);".format(t)),
 		Function("__float__", [Argument(constRef, "this_")], "return librapid::internal::traits<librapid::{}>::cast<double>(this_);".format(t)),
 
-		Function("str", [Argument(constRef, "this_"), Argument("int64_t", "base", "10")], "return lrc::str(this_, {-1, base, false});"),
+		Function("str", [Argument(constRef, "this_"), Argument("int8_t", "base", "10")], "return lrc::str(this_, {-1, base, false});"),
 		Function("__str__", [Argument(constRef, "this_")], "return lrc::str(this_, {-1, 10, false});"),
 		Function("__repr__", [Argument(constRef, "this_")], "return \"librapid::{}(\\\"\" + lrc::str(this_, {{-1, 10, false}}) + \"\\\")\";".format(t)),
 	]
